fixed bug, still not refactored
