+++ src/scripts/events/initEvents.js
@@ -3,11 +3,13 @@ const showEventView = require("./eventView")
 const activateEventFormButtons = require("./eventFormButtons")
 const createEventForm = require("./newEventForm")
 const createEditEventForm = require("./createEditEventForm")
+const loadEvents = require("./loadEvents");
+
 
 const initializeEvents = () => {
     showEventView()
+    loadEvents();
     createEventForm()
-    activateEventFormButtons()
 }
 
 module.exports = initializeEvents
